requirements.txt
# Beacon - Milestone 1 & 2 Dependencies
# Python 3.11+

# Core Dependencies
pydantic>=2.5.0
pyyaml>=6.0.1
structlog>=24.1.0

# Safety Service - Regex & Semantic
google-re2>=1.1
sentence-transformers>=2.2.2
# onnxruntime>=1.16.0  # Not available for Python 3.14 yet
numpy>=1.24.0
torch>=2.1.0

# Reasoning Module - Mistral-7B
transformers>=4.36.0
accelerate>=0.25.0

# Conversation Agent - Native Llama Integration
llama-cpp-python>=0.2.0
nvidia-ml-py>=12.535.0  # For VRAM detection

# LangChain (kept for compatibility, but not used in native mode)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20

# Testing & Evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0

# Utilities
python-dotenv>=1.0.0
